# routes.php
<?php

use Cms\Classes\Controller;
use Illuminate\Http\Request;
use KodZero\POSMall\Classes\Api\ApiRouteAccessGuard;
use KodZero\POSMall\Classes\Api\ApiTokenGuard;
use KodZero\POSMall\Classes\Api\CartApiService;
use KodZero\POSMall\Classes\Api\CatalogApiService;
use KodZero\POSMall\Classes\Api\CheckoutApiService;
use KodZero\POSMall\Classes\Api\CommerceContext;
use KodZero\POSMall\Classes\Api\CustomerApiService;
use KodZero\POSMall\Classes\Api\DiscoveryApiService;
use KodZero\POSMall\Classes\Api\FavoriteListApiService;
use KodZero\POSMall\Classes\Api\GraphQL\GraphqlApiService;
use KodZero\POSMall\Classes\Api\JsonResponder;
use KodZero\POSMall\Classes\Api\JsonResponseMarker;
use KodZero\POSMall\Classes\Api\OrderApiService;
use KodZero\POSMall\Classes\Api\PaymentLinkService;
use KodZero\POSMall\Classes\Api\ReviewsApiService;
use KodZero\POSMall\Classes\Api\ServicesApiService;
use KodZero\POSMall\Classes\Api\TaxPreviewApiService;
use KodZero\POSMall\Classes\Feeds\GoogleMerchantFeed;
use KodZero\POSMall\Classes\Payments\Webhooks\StripeHostedCheckoutWebhook;
use KodZero\POSMall\Models\ApiSettings;
use KodZero\POSMall\Models\FeedSettings;

// Send virtual product files to the user's browser.
$virtualProductDownloadHandler = '\KodZero\POSMall\Classes\Downloads\VirtualProductFileDownload@handle';

Route::get('/posmall/download/{key}/{idx?}', $virtualProductDownloadHandler)
    ->middleware('web');

// Handles Responses generated by a PaymentProvider. It just renders the contents of
// the posmall.checkout.response session key. This way a PaymentProvider may render a HTML
// form to execute custom POST requests to a PaymentGateway.
//
// @see \KodZero\POSMall\Classes\Payments\PaymentRedirector::handlePaymentResult
$paymentResponseHandler = function () {
    $data = session()->pull('posmall.checkout.response');

    if (! $data) {
        return response('Missing data.', 404);
    }

    return response($data, 200);
};

Route::get('/posmall/checkout/response', $paymentResponseHandler)
    ->middleware('web');

// Renders the Google Merchant integration feed. It outputs all
// of the store's products as an XML document, which is indexed
// by Google on a regular basis.
Route::get('/feeds/google-merchant/{key}', function ($key) {
    $useFeed = FeedSettings::get('google_merchant_enabled');

    if (! $useFeed) {
        return (new Controller())->run('404');
    }

    $settingsKey = FeedSettings::get('google_merchant_key');

    if ($key !== $settingsKey) {
        return response('Auth failed', 401);
    }

    $feed = new GoogleMerchantFeed(
        request()->get('locale')
    );

    return response($feed->build(), 200, ['Content-Type' => 'application/xml']);
});

// Handles Stripe Checkout webhook events (backup payment verification).
// This ensures payment confirmation happens even if the user doesn't complete
// the browser redirect after payment.
$stripeWebhookHandler = function () {
    return (new StripeHostedCheckoutWebhook())->handle();
};

Route::post('/posmall/webhooks/stripe', $stripeWebhookHandler);

Route::get('/posmall/pay/{token}', function (string $token) {
    $resolved = app(PaymentLinkService::class)->resolve($token);

    if ($resolved instanceof JsonResponseMarker) {
        return response($resolved->message, $resolved->status);
    }

    session()->put('posmall.processing_order.id', $resolved->hash_id);
    session()->put('posmall.checkout.flow', 'payment-link');

    return redirect()->to('/posmall/payment-link/' . $resolved->hash_id . '?payment_link=1');
})->where('token', 'SMSPay_[A-Fa-f0-9]{32}')->middleware('web');

$posmallApi = function (array $scopes, callable $handler) {
    return function (Request $request, ...$routeParameters) use ($scopes, $handler) {
        $responder = app(JsonResponder::class);
        $auth = app(ApiRouteAccessGuard::class)->authenticate($request, $scopes)
            ?? app(ApiTokenGuard::class)->authenticate($request, $scopes);

        if ($auth instanceof JsonResponseMarker) {
            return $responder->error($auth->code, $auth->message, $auth->status, $auth->meta);
        }

        $context = CommerceContext::fromRequest($auth, $request);

        if ($context instanceof JsonResponseMarker) {
            return $responder->error($context->code, $context->message, $context->status, $context->meta);
        }

        try {
            return $handler($request, $responder, $context, ...$routeParameters);
        } catch (Throwable $exception) {
            return $responder->exception($exception);
        }
    };
};

$posmallApiInput = fn (Request $request): array => array_replace(
    $request->query(),
    $request->request->all(),
    is_array($request->json()->all()) ? $request->json()->all() : []
);

Route::post('/posmall/api/graphql', $posmallApi([], function (Request $request, JsonResponder $responder, CommerceContext $context) {
    if (!ApiSettings::graphqlEnabled()) {
        return $responder->error('graphql_disabled', 'POSMall GraphQL is disabled.', 403);
    }

    $response = app(GraphqlApiService::class)->execute($request, $context);

    return response()->json($response['body'], $response['status']);
}));

Route::prefix('/posmall/api/v1')->group(function () use ($posmallApi, $posmallApiInput) {
    Route::get('/status', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) {
        return $responder->ok([
            'status' => 'ok',
            'version' => 'v1',
            'context' => [
                'vendor' => optional($context->vendor)->only(['id', 'name', 'slug']),
                'channel' => optional($context->channel)->only(['id', 'name', 'slug', 'type']),
                'warehouse' => optional($context->warehouse)->only(['id', 'name', 'slug', 'type']),
            ],
        ]);
    }));

    Route::get('/categories', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder) {
        return $responder->ok([
            'categories' => app(CatalogApiService::class)->categories(),
        ]);
    }));

    Route::get('/brands', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder) {
        return $responder->ok(app(DiscoveryApiService::class)->brands($request->query()));
    }));

    Route::get('/brands/{slug}/products', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $slug) {
        return $responder->ok(app(DiscoveryApiService::class)->brandProducts($slug, $request->query(), $context));
    }))->where('slug', '[A-Za-z0-9\/:_\-\*\[\]\+\?\|]+');

    Route::get('/brands/{slug}', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $slug) {
        return $responder->ok(app(DiscoveryApiService::class)->brandBySlug($slug));
    }))->where('slug', '[A-Za-z0-9\/:_\-\*\[\]\+\?\|]+');

    Route::get('/properties', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder) {
        return $responder->ok(app(DiscoveryApiService::class)->properties($request->query()));
    }));

    Route::get('/vendors', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) {
        return $responder->ok(app(DiscoveryApiService::class)->vendors($request->query() + ['context' => $context]));
    }));

    Route::get('/vendors/{slug}/products', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $slug) {
        return $responder->ok(app(DiscoveryApiService::class)->vendorProducts($slug, $request->query(), $context));
    }))->where('slug', '[A-Za-z0-9:_\-]+');

    Route::get('/vendors/{slug}', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $slug) {
        return $responder->ok(app(DiscoveryApiService::class)->vendorBySlug($slug, $context));
    }))->where('slug', '[A-Za-z0-9:_\-]+');

    Route::get('/channels', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) {
        return $responder->ok(app(DiscoveryApiService::class)->channels($request->query() + ['context' => $context]));
    }));

    Route::get('/channels/{slug}', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $slug) {
        return $responder->ok(app(DiscoveryApiService::class)->channelBySlug($slug, $context));
    }))->where('slug', '[A-Za-z0-9:_\-]+');

    Route::get('/warehouses', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) {
        return $responder->ok(app(DiscoveryApiService::class)->warehouses($request->query() + ['context' => $context]));
    }));

    Route::get('/warehouses/{slug}/stock', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $slug) {
        return $responder->ok(app(DiscoveryApiService::class)->warehouseStock($slug, $request->query(), $context));
    }))->where('slug', '[A-Za-z0-9:_\-]+');

    Route::get('/warehouses/{slug}', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $slug) {
        return $responder->ok(app(DiscoveryApiService::class)->warehouseBySlug($slug, $context));
    }))->where('slug', '[A-Za-z0-9:_\-]+');

    Route::get('/products', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) {
        return $responder->ok(app(CatalogApiService::class)->list($request->query(), $context));
    }));

    Route::get('/products/{slug}/reviews', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $slug) {
        return $responder->ok(app(ReviewsApiService::class)->list($slug, $request->query()));
    }))->where('slug', '[A-Za-z0-9\/:_\-\*\[\]\+\?\|]+');

    Route::post('/products/{slug}/reviews', $posmallApi(['reviews:write'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $slug) use ($posmallApiInput) {
        return $responder->ok(app(ReviewsApiService::class)->create($slug, $posmallApiInput($request), $context), 201);
    }))->where('slug', '[A-Za-z0-9\/:_\-\*\[\]\+\?\|]+');

    Route::get('/products/{slug}', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $slug) {
        return $responder->ok(app(CatalogApiService::class)->product($slug, $context));
    }))->where('slug', '[A-Za-z0-9\/:_\-\*\[\]\+\?\|]+');

    Route::get('/services', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) {
        return $responder->ok(app(ServicesApiService::class)->list($request->query(), $context));
    }));

    Route::get('/services/{code}', $posmallApi(['catalog:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $code) {
        return $responder->ok(app(ServicesApiService::class)->detail($code, $context));
    }))->where('code', '[A-Za-z0-9:_\-]+');

    Route::post('/services/{code}/cart', $posmallApi(['cart:write'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $code) use ($posmallApiInput) {
        $input = $posmallApiInput($request);
        $service = app(ServicesApiService::class)->detail($code, $context)['service'];
        $input['service_id'] = $service['id'];

        return $responder->ok(app(ServicesApiService::class)->addToCart($input, $context), 201);
    }))->where('code', '[A-Za-z0-9:_\-]+');

    Route::get('/cart', $posmallApi(['cart:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(CartApiService::class)->get($posmallApiInput($request), $context));
    }));

    Route::get('/cart/shipping-methods', $posmallApi(['cart:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(CartApiService::class)->shippingMethods($posmallApiInput($request), $context));
    }));

    Route::get('/cart/payment-methods', $posmallApi(['cart:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(CartApiService::class)->paymentMethods($posmallApiInput($request), $context));
    }));

    Route::get('/cart/tax-preview', $posmallApi(['cart:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(TaxPreviewApiService::class)->cart($posmallApiInput($request), $context));
    }));

    Route::post('/cart/items', $posmallApi(['cart:write'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(CartApiService::class)->addItem($posmallApiInput($request), $context), 201);
    }));

    Route::patch('/cart/items/{item}', $posmallApi(['cart:write'], function (Request $request, JsonResponder $responder, CommerceContext $context, int $item) use ($posmallApiInput) {
        return $responder->ok(app(CartApiService::class)->setQuantity($posmallApiInput($request), $item, $context));
    }))->whereNumber('item');

    Route::delete('/cart/items/{item}', $posmallApi(['cart:write'], function (Request $request, JsonResponder $responder, CommerceContext $context, int $item) use ($posmallApiInput) {
        return $responder->ok(app(CartApiService::class)->removeItem($posmallApiInput($request), $item, $context));
    }))->whereNumber('item');

    Route::post('/cart/discounts', $posmallApi(['cart:write'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(CartApiService::class)->applyDiscount($posmallApiInput($request), $context));
    }));

    Route::post('/cart/shipping-method', $posmallApi(['cart:write'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(CartApiService::class)->setShippingMethod($posmallApiInput($request), $context));
    }));

    Route::post('/cart/payment-method', $posmallApi(['cart:write'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(CartApiService::class)->setPaymentMethod($posmallApiInput($request), $context));
    }));

    Route::post('/cart/address', $posmallApi(['cart:write'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(CartApiService::class)->setAddress($posmallApiInput($request), $context));
    }));

    Route::post('/checkout/orders', $posmallApi(['checkout:write'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(CheckoutApiService::class)->createOrder($posmallApiInput($request), $context), 201);
    }));

    Route::get('/orders', $posmallApi(['orders:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(OrderApiService::class)->list($posmallApiInput($request), $context));
    }));

    Route::get('/orders/{hash}/status', $posmallApi(['orders:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $hash) use ($posmallApiInput) {
        return $responder->ok(app(OrderApiService::class)->status($hash, $posmallApiInput($request), $context));
    }));

    Route::get('/orders/{hash}', $posmallApi(['orders:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, string $hash) use ($posmallApiInput) {
        return $responder->ok(app(OrderApiService::class)->detail($hash, $posmallApiInput($request), $context));
    }));

    Route::get('/customers/{customer}', $posmallApi(['customers:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, int $customer) {
        return $responder->ok(app(CustomerApiService::class)->get($customer, $context));
    }))->whereNumber('customer');

    Route::get('/customers/{customer}/addresses', $posmallApi(['customers:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, int $customer) {
        return $responder->ok(app(CustomerApiService::class)->addresses($customer, $context));
    }))->whereNumber('customer');

    Route::get('/customers/{customer}/orders', $posmallApi(['orders:read'], function (Request $request, JsonResponder $responder, CommerceContext $context, int $customer) use ($posmallApiInput) {
        return $responder->ok(app(OrderApiService::class)->list($posmallApiInput($request), $context, $customer));
    }))->whereNumber('customer');

    Route::post('/customers/{customer}/addresses', $posmallApi(['customers:write'], function (Request $request, JsonResponder $responder, CommerceContext $context, int $customer) use ($posmallApiInput) {
        return $responder->ok(app(CustomerApiService::class)->createAddress($customer, $posmallApiInput($request), $context), 201);
    }))->whereNumber('customer');

    Route::patch('/customers/{customer}/addresses/{address}', $posmallApi(['customers:write'], function (Request $request, JsonResponder $responder, CommerceContext $context, int $customer, int $address) use ($posmallApiInput) {
        return $responder->ok(app(CustomerApiService::class)->updateAddress($customer, $address, $posmallApiInput($request), $context));
    }))->whereNumber('customer')->whereNumber('address');

    Route::get('/favorite-lists', $posmallApi(['favorites:read'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(FavoriteListApiService::class)->list($posmallApiInput($request), $context));
    }));

    Route::post('/favorite-lists', $posmallApi(['favorites:write'], function (Request $request, JsonResponder $responder, CommerceContext $context) use ($posmallApiInput) {
        return $responder->ok(app(FavoriteListApiService::class)->create($posmallApiInput($request), $context), 201);
    }));

    Route::post('/favorite-lists/{list}/items', $posmallApi(['favorites:write'], function (Request $request, JsonResponder $responder, CommerceContext $context, int $list) use ($posmallApiInput) {
        return $responder->ok(app(FavoriteListApiService::class)->addItem($posmallApiInput($request), $list, $context), 201);
    }))->whereNumber('list');

    Route::delete('/favorite-lists/{list}/items/{item}', $posmallApi(['favorites:write'], function (Request $request, JsonResponder $responder, CommerceContext $context, int $list, int $item) use ($posmallApiInput) {
        return $responder->ok(app(FavoriteListApiService::class)->removeItem($posmallApiInput($request), $list, $item, $context));
    }))->whereNumber('list')->whereNumber('item');
});